[fix] Edge tool preconditions are not properly supported
At the moment, the edge tool preconditions are evaluated as soon as the
node that contains the edge tool is selected. Since no edges are built
at this stage, edge variables are not instantiated and accessing these
variables raises errors.
A proposal (https://github.com/eclipse-sirius/sirius-web/issues/2930) is
made to evaluate preconditions when the edge target is selected. That
way, the edge variables will be set and could be used in services.

The fix consists to remove the precondition in the edge tool to avoid
errors.

Bug: https://github.com/eclipse-syson/syson/issues/169
Signed-off-by: Jerome Gout <jerome.gout@obeosoft.com>