[695] Handle representation of Documentation node on graphical views
It is now possible to represent Documentation element as a node on each
available graphical views. User can now Drag&Drop a Documentation from
the explorer view to the Diagram to show it.
Each Documentation node is linked to its annotated element
representation by a dash line. The Documentation node can only appear
if its annotated element is represented on the diagram. User can direct
edit the body of the Documentation.

Bug: https://github.com/eclipse-syson/syson/issues/695
Signed-off-by: Jessy Mallet <jessy.mallet@obeo.fr>