[696] Handle representation of Comment node on graphical views
It is now possible to create and represent Comment element as a node on
each available graphical views. User can create a Comment element from
the explorer view to the Diagram to show it.
User can now Drag&Drop a Comment from any element on diagrams except on
Comment or Documentation.

Each Comment node is linked to its annotated element
representation by a dash line. The Comment node can only appear
if its annotated element is represented on the diagram. User can direct
edit the body of the Comment.

Bug: https://github.com/eclipse-syson/syson/issues/696
Signed-off-by: Jessy Mallet <jessy.mallet@obeo.fr>