[1018] Rework library loading to be extensible and re-usable
* Move our library loading code into SysONLibraryLoader, which
downstream applications may re-use to load their own libraries from the
classpath.
* Also make the library loading unit tests rely on them.
* Make internal methods of SysMLStandardLibrariesConfiguration protected
instead of private so it can be easily extended/redefined.
* The list of EPackages to register, and of libraries to load, are now
defined in their own protected method so as to be easily redefined.
* Move KerML/SysML standard libraries loading out of the constructor of
the @Configuration. Now the standard libraries get loaded when they are
needed for the first time. Downstream applications may choose this
moment to be at application start, or, by default, at the latest time
possible, that is, when first opening a SysML project.
* Rename 'SysMLStandardLibrariesConfiguration' into
'SysONDefaultLibrariesConfiguration' to reflect the fact that default
libraries may include non-standard libraries and that it is up to the
application to determine which libraries are the default ones.
* Update changelog and release notes accordingly.

Bug: https://github.com/eclipse-syson/syson/issues/1018
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>