[1078] Add support for publishing SysML projects as libraries
Note that there is still no way to use published libraries in SysON, as
this will be the focus of a later feature.

* Backend
** Add API and implementation to handle the publication of the SysML
contents of a Project as a 'library'.
Note: Sirius Web 'libraries' and SysML 'libraries' have no particular
link besides being namesakes.
** Add @Service implementation for SysON's omnibox command provider.
This implementation only provides 1 command that is used to publish the
SysML contents of a project.

* Frontend
** Create a command for the Omnibox that is meant for the publication of
the SysML contents of a project as a library.
** Add this command to the default extension registry of the SysON
application.
** Un-hide in the 'help' menu (top right corner) the link to the list of
published libraries (i.e. page '<baseUrl>/libraries/'). This entry was
hidden previously because SysON did not make use of the publication
mechanism.

* Add integration tests for ensuring the handler that publishes a
project as a library works as intended.
* Update release notes, changelog and user manual accordingly.

Bug: https://github.com/eclipse-syson/syson/issues/1078
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>