[1118] Add support for updating version of library dependencies
Users can now upgrade or downgrade the version of a library declared as
dependency to one of their projects by using the dedicated action from
the contextual menu in the 'Explorer' view.

* Backend
** Tweak publication to _not_ include the library version in the
resource ID of the published libraries, this way we can maintain
identity of the resources across versions.
** Add SysON-specific implementation of
ITreeItemContextMenuEntryProvider to make contextual menu entry 'Update
library' appear in the 'Explorer' view.

* Frontend
** Add tree item context menu contributions to implement the 'Update
library' action

* Tests
** Add integration tests

* Documentation
** Update user documentation to include a section about 'Library
update'.
** Update changelog
** Update release notes

Bug: https://github.com/eclipse-syson/syson/issues/1118
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>