[1452] Restrict "New objects from text" action to SysML model elements
* Move contextual menu action "New objects from text" to the backend.
Previously this command used the "kind" property of TreeItems to
determine on which elements it was available. With Sirius Web
deprecating this property, moving the action to the backend was the
identified migration path. This should help the maintainability and
testability of the action.
* Add in the SysON explorer services, a service to determine on which
elements this action is available. The default implementation only
enables it on SysML model elements.
* Update changelog and release notes accordingly.

Bug: https://github.com/eclipse-syson/syson/issues/1452
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>